+++ R/utils.R
@@ -215,30 +215,61 @@ check_function2 <- function(x,
   }
 
   actual_args <- fn_fmls_names(f) %||% character()
-  if (identical(actual_args, expected_args)) {
+  missing_args <- setdiff(expected_args, actual_args)
+  if (is_empty(missing_args)) {
     return(invisible(NULL))
   }
 
   n_expected_args <- length(expected_args)
   n_actual_args <- length(actual_args)
 
-  if (n_expected_args == 0) {
-    cli::cli_abort(
-      "{.arg {arg}} must have no arguments, not {n_actual_args} argument{?s}.",
-      call = call,
-      arg = arg
-    )
-  }
-
   if (n_actual_args == 0) {
-    arg_info <- "instead of no arguments"
+    arg_info <- "instead it has no arguments"
   } else {
-    arg_info <- "not {.arg {actual_args}}"
+    arg_info <- paste0("it currently has {.arg {actual_args}}")
   }
 
   cli::cli_abort(
-    paste0("{.arg {arg}} must have the {cli::qty(n_expected_args)}argument{?s} {.arg {expected_args}}, ", arg_info, "."),
+    paste0("{.arg {arg}} must have the {cli::qty(n_expected_args)}argument{?s} {.arg {expected_args}}; ", arg_info, "."),
     call = call,
     arg = arg
   )
+}
+
+# This is inspired by the C interface of `cli_progress_bar()` which has just
+# 2 arguments: `total` and `config`
+create_progress_bar <- function(total,
+                                name,
+                                config,
+                                env = caller_env(),
+                                config_arg = caller_arg(config),
+                                error_call = caller_env()) {
+  if (is_false(config)) {
+    return()
+  }
+
+  if (is.null(config) || is_bool(config)) {
+    args <- list()
+  } else if (is_scalar_character(config)) {
+    args <- list(name = config)
+  } else if (is.list(config)) {
+    args <- config
+  } else {
+    stop_input_type(
+      config,
+      what = c("a bool", "a string", "a list"),
+      arg = config_arg,
+      call = error_call
+    )
+  }
+
+  args$name <- args$name %||% name
+  # Can be removed if https://github.com/r-lib/cli/issues/630 is fixed
+  if (is.infinite(total)) {
+    total <- NA
+  }
+  args$total <- total
+  args$.envir <- env
+
+  exec(cli::cli_progress_bar, !!!args)
 }