+++ NEWS.md
@@ -1,5 +1,21 @@
 # httr2 (development version)
 
+* `req_oauth_device()` now takes a `auth_url` parameter making it usable 
+  (#331, @taerwin).
+
+* OAuth errors containing a url now correctly display that URL 
+  (instead of just uri).
+
+* New `req_paginate()` and `paginate_req_perform()` to easily request paginated
+  APIs (@mgirlich, #8).
+
+* New `resp_check_content_type()` to check response content types (#190).
+  `resp_body_json()` and friends give better errors if no `Content-Type`
+  header is present in the response (#284).
+
+* `req_body_json()` gains custom content `type` argument
+  and respects custom content-type set in header (@mgirlich, #189).
+
 * `oauth_flow_auth_code_listen()` now takes a single `redirect_uri` argument
   instead of separate `host_ip` and `port` arguments. This is a breaking change
   but I don't expect anyone to call this function directly (which was confirmed